+++ packages/react-databrowser/src/components/databrowser/hooks/useFetchSingleDataByKey/fetch-data-types.ts
@@ -5,6 +5,7 @@ import {
   INITIAL_CURSOR_NUM,
   toJsonStringifiable,
   transformArray,
+  transformHash,
   transformStream,
 } from "./utils";
 
@@ -47,9 +48,8 @@ export const fetchDataOfType = {
     if (index === cursorStack.current.length - 1) {
       cursorStack.current.push(nextCursor);
     }
-
-    const content = transformArray(hashValues);
-    return { content: content, type: "hash", memory: roughSizeOfObject(content) } satisfies {
+    const content = transformHash(hashValues);
+    return { content, type: "hash", memory: roughSizeOfObject(content) } satisfies {
       content: ContentValue[];
       type: "hash";
       memory: number;